refactor: update prop types and imports
