+++ src/Donations/resources/DonationDetails/app/components/PaymentInformation/DonationMethod/index.tsx
@@ -1,9 +1,12 @@
-import {DonationMethodProps} from './types';
-
 /**
  *
  * @unreleased
  */
+
+export type DonationMethodProps = {
+    gatewayLabel: string;
+};
+
 export default function DonationMethod({gatewayLabel}: DonationMethodProps) {
     return <span>{gatewayLabel}</span>;
 }
