--- src/Donations/resources/DonationDetails/app/components/PaymentInformation/DonationMethod/types.ts
@@ -1,3 +1,0 @@
-export type DonationMethodProps = {
-    gatewayLabel: string;
-}
