--- src/Donations/resources/DonationDetails/app/components/PaymentInformation/DonationType/types.ts
@@ -1,3 +1,0 @@
-export type DonationTypeProps = {
-    donationType: 'single' | 'subscription' | string | null;
-};
