+++ src/Donations/resources/DonationDetails/app/components/PaymentInformation/Fields/Amount/index.tsx
@@ -4,6 +4,7 @@ import {CurrencyField} from '../Field';
 import {useFormContext, useWatch} from 'react-hook-form';
 import {formatCurrency} from '../../../../utilities/formatter';
 import {__} from '@wordpress/i18n';
+
 import {CurrencyInputField} from '@givewp/components/AdminUI/FormElements';
 import Button from '@givewp/components/AdminUI/Button';
 import NoticeInformationIcon from '@givewp/components/AdminUI/Icons/NoticeInformationIcon';
