--- src/Donations/resources/DonationDetails/app/components/PaymentInformation/Legend/types.ts
@@ -1,4 +1,0 @@
-export type LegendProps = {
-    title: string;
-    donationType: string;
-};
