--- src/Views/Components/AdminUI/ActionMenu/types.ts
@@ -1,4 +1,0 @@
-export type ActionMenuProps = {
-    menuConfig: any;
-    toggle: any;
-}
