+++ src/Views/Components/AdminUI/Button/index.tsx
@@ -2,14 +2,23 @@ import React from 'react';
 
 import cx from 'classnames';
 
-import {ButtonProps} from './types';
-
 import styles from './style.module.scss';
 
 /**
  *
  * @unreleased
  */
+
+export type ButtonProps = {
+    variant?: 'primary' | 'secondary' | 'danger';
+    size?: 'small' | 'large';
+    type?: 'button' | 'reset' | 'submit';
+    children: React.ReactNode;
+
+    onClick?: React.MouseEventHandler<HTMLButtonElement>;
+    disabled?: boolean;
+    classname?: 'string';
+};
 
 const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
     ({children, type = 'button', classname, variant = 'primary', size = 'normal', disabled, ...props}, ref) => (
