--- src/Views/Components/AdminUI/FormElements/types.ts
@@ -1,26 +1,0 @@
-export type FormElementProps = {
-    children: React.ReactNode;
-    onSubmit: React.FormEventHandler<HTMLFormElement>;
-    id: string;
-};
-
-export type TextInputFieldProps = {
-    name: string;
-    type: string;
-    placeholder: string;
-    label: string;
-};
-
-export type CurrencyInputFieldProps = {
-    name: string;
-    type: string;
-    placeholder: string;
-    label: string;
-    currency: string;
-    defaultValue;
-
-    handleCurrencyChange;
-    control;
-
-    value?: any;
-};
