--- src/Views/Components/AdminUI/SearchSelector/types.ts
@@ -1,9 +1,0 @@
-export type SearchSelector = {
-    options: Array<{
-        value: number;
-        label: string;
-    }>;
-    defaultLabel?: string;
-    name: string;
-    placeholder: string;
-};
