--- src/Views/Components/AdminUI/StatusSelector/types.ts
@@ -1,6 +1,0 @@
-export type StatusSelector = {
-    options: Array<{
-        value: string;
-        label: string;
-    }>;
-};
