--- src/Views/Components/AdminUI/Toast/types.ts
@@ -1,6 +1,0 @@
-export type ToastProps = {
-    resultType: 'success' | 'error' | null
-    resultMessage: string
-    closeMessage:() => void
-    showMessage: boolean
-};
